Avoid NPE on misconfigured styles; fixes GEOS-4002
Before, the CSS demo page selected a style and attempted to load it
without checking that the file existed.  Now, the demo page checks it
and uses an error message in place of the SLD text if it cannot be
loaded.  Also, now the default style for the layer is loaded if none
is specified as a query parameter.

git-svn-id: https://svn.codehaus.org/geoserver/branches/2.0.x/src@14482 ef1d6e69-97e5-0310-af46-8a06194da32a